+++ Configuration/TCA/Overrides/tt_content.php
@@ -3,6 +3,6 @@
 defined('TYPO3') or die();
 
 call_user_func(function () {
-    $obj = new \Slavlee\CustomPackage\Bootstrap\TCA\SysTemplate();
+    $obj = new \Slavlee\CustomPackage\Bootstrap\TCA\TtContent();
     $obj->invoke();
 });