+++ CHANGES.md
@@ -1,3 +1,11 @@
+## Version 0.0.7
+- [FEATURE] Add example on how to configure plugin
+- [FEATURE] Add example ActionController, Domain Model and Repository
+- [FEATURE] Add example Extbase/Persistence/Classes
+- [FEATURE] Add example for QueryResultPagination
+- [FEATURE] Add Base->registerFlexform
+- [FEATURE] Add example FileUploadService
+
 ## Version 0.0.6
 - [FEATURE] Add example language file
 - [FEATURE] Add lib.dynamicContent to flexible read content elements